# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/blogs:
    get:
      tags:
        - Blogs
      summary: List Blogs
      description: Get a paginated list of Blogs with optional filtering and sorting
      operationId: listBlogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Blogs
      summary: Create Blog
      description: Create a new Blog
      operationId: createBlog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Blog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/blog_audit_logs:
    get:
      tags:
        - BlogAuditLogs
      summary: List BlogAuditLogs
      description: Get a paginated list of BlogAuditLogs with optional filtering and sorting
      operationId: listBlogAuditLogs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogAuditLogList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - BlogAuditLogs
      summary: Create BlogAuditLog
      description: Create a new BlogAuditLog
      operationId: createBlogAuditLog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBlogAuditLogRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogAuditLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Blog:
      type: object
      required:
        - id
        - website_id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        website_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        subtitle:
          nullable: true
          type: string
          maxLength: 500
        description:
          nullable: true
          type: string
        archived_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreateBlogRequest:
      type: object
      required:
        - website_id
        - name
        - metadata
      properties:
        website_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 120
        subtitle:
          nullable: true
          type: string
          maxLength: 500
        description:
          nullable: true
          type: string
        archived_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    BlogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Blog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    BlogAuditLog:
      type: object
      required:
        - id
        - event
        - occurred_at
      properties:
        id:
          type: string
          format: uuid
        event:
          type: string
        actor:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
          maxLength: 120
        subject_id:
          nullable: true
          type: string
          format: uuid
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
...
    CreateBlogAuditLogRequest:
      type: object
      required:
        - event
        - occurred_at
      properties:
        event:
          type: string
        actor:
          nullable: true
          type: string
          format: uuid
        subject_type:
          nullable: true
          type: string
          maxLength: 120
        subject_id:
          nullable: true
          type: string
          format: uuid
        detail:
          nullable: true
          type: object
        occurred_at:
          type: string
          format: date-time
...
    BlogAuditLogList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BlogAuditLog'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []